CategoryAdapter moved to a separate class
